Move all line handle movement logic to one file
Fix guides for lines.
